batch: Group saved ownership modules
The batch package has bounded matching and realization dependencies while
saved ownership remains spread across many flat model, translation,
metadata, and selection modules.

Ownership changes require contributors to distinguish those files from
unrelated merge, storage, source, and display workflows in the same
directory.

This commit moves the connected ownership model and its transformations
beneath ownership, shortens repeated filenames, and leaves cross-package
update orchestration at the batch root.

Later commits will migrate merge, source, storage, and display callers
before the temporary ownership imports are removed.
