+++ src/git_stage_batch/batch/ownership/__init__.py
@@ -1,0 +1,6 @@
+"""Saved line ownership values and transformations."""
+
+from .model import BatchOwnership, ResolvedBatchOwnership
+
+
+__all__ = ["BatchOwnership", "ResolvedBatchOwnership"]
