+++ src/git_stage_batch/batch/ownership_detachment.py
@@ -1,56 +1,8 @@
-"""Detached ownership copies for use outside acquisition scopes."""
+"""Compatibility import for batch ownership detachment."""
 
-from __future__ import annotations
+import sys as _sys
 
-from ..core.buffer import LineBuffer
-from .absence_content import copy_absence_content as _copy_absence_content
-from .ownership_acquisition import AcquiredBatchOwnership
-from .ownership_absence_claims import AbsenceClaim
-from .ownership import BatchOwnership
-from .ownership_claims import PresenceClaim
-from .ownership_replacement_units import ReplacementUnit
+from .ownership import detachment as _implementation
 
 
-def acquire_detached_batch_ownership(
-    ownership: BatchOwnership,
-) -> AcquiredBatchOwnership[BatchOwnership]:
-    """Acquire an ownership copy with independent absence content buffers."""
-    buffers: list[LineBuffer] = []
-    deletions: list[AbsenceClaim] = []
-    try:
-        for deletion in ownership.deletions:
-            content_lines = _copy_absence_content(deletion.content_lines)
-            buffers.append(content_lines)
-            deletions.append(
-                AbsenceClaim(
-                    anchor_line=deletion.anchor_line,
-                    content_lines=content_lines,
-                    baseline_reference=deletion.baseline_reference,
-                )
-            )
-    except Exception:
-        for buffer in buffers:
-            buffer.close()
-        raise
-
-    return AcquiredBatchOwnership(
-        ownership=BatchOwnership(
-            presence_claims=[
-                PresenceClaim(
-                    source_lines=claim.source_lines[:],
-                    baseline_references=dict(claim.baseline_references),
-                )
-                for claim in ownership.presence_claims
-            ],
-            deletions=deletions,
-            replacement_units=[
-                ReplacementUnit(
-                    presence_lines=unit.presence_lines[:],
-                    deletion_indices=unit.deletion_indices[:],
-                    origin=unit.origin,
-                )
-                for unit in ownership.replacement_units
-            ],
-        ),
-        buffers=buffers,
-    )
+_sys.modules[__name__] = _implementation
